requirements.txt
py-slippi
tqdm
pandas
